requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.54.4
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.0
typing-extensions==4.12.2
uvicorn==0.32.0
